pyproject.toml
[project]
name = "main-diary-be-15"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aerich[toml]>=0.9.2",
    "aiomysql>=0.3.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "email-validator>=2.3.0",
    "fastapi>=0.124.0",
    "greenlet>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.8",
    "sqlalchemy>=2.0.44",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
]

[tool.aerich]
tortoise_orm = "app.db.base.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
